Regression in gcc-arm 8 release series
Fix an issue with using `ORIGIN()` or `LENGTH()` expressions in MEMORY
region declarations- this is a regression in the arm-gcc-8 release
series:
https://sourceware.org/bugzilla/show_bug.cgi?id=24289

Switch to using a variable and a more hard-coded expression that doesn't
fail to build using those compilers.

Verified with 8-2019-q3-update.

Also fix an error in the arithmetic used to reserve the interrupt
regions (it should have been eg `(16 + 240) * 4` instead of
`(16 * 240) *4` 😬.

Fixes #5.