🚿 mark .heap section as NOLOAD not COPY
COPY sections are omitted from tools like arm-none-eabi-size / readelf .
Use NOLOAD for the .heap section instead; it still won't show up in
those tools, since the .heap section is not placed (no symbols are in
it). The .heap section is used only for the start/end addresses used by
sbrk in newlib.

Also add the `--print-memory-usage` option to the link stage so we can
see section usage stats when linking.